requirements.txt
langchain==1.0.8
langchain-core==1.1.0
langchain-google-genai==3.1.0
python-dotenv
requests
uvicorn
fastapi
